pyproject.toml
[project]
name = "vakit-pi"
version = "0.1.0"
description = "Raspberry Pi için Namaz Vakti ve Ezan Uygulaması"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Vakit-Pi Developers" }
]
keywords = ["prayer", "adhan", "namaz", "ezan", "raspberry-pi", "islam"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Religion",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # Prayer time calculation
    "islam>=0.6.0",
    "timezonefinder>=6.5.0",
    
    # Scheduling
    "apscheduler>=3.10.0",
    
    # Data handling
    "pydantic>=2.10.0",
    "aiofiles>=24.1.0",
    
    # Internationalization
    "babel>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-aiofiles>=24.1.0",
]

[project.scripts]
vakit-pi = "vakit_pi.cli:main"

[project.urls]
Homepage = "https://github.com/vakit-pi/vakit-pi"
Repository = "https://github.com/vakit-pi/vakit-pi"
Issues = "https://github.com/vakit-pi/vakit-pi/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vakit_pi"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

# ============== Ruff Configuration ==============
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "RUF001", # ambiguous unicode characters in strings (Turkish chars)
    "RUF002", # ambiguous unicode in docstrings (Turkish chars)
    "RUF003", # ambiguous unicode in comments (Turkish chars)
]

[tool.ruff.lint.isort]
known-first-party = ["vakit_pi"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============== MyPy Configuration ==============
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pyIslam.*",
    "apscheduler.*",
    "timezonefinder.*",
]
ignore_missing_imports = true

# ============== Pytest Configuration ==============
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/vakit_pi"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
